requirements.txt
asgiref==3.8.1
Django==5.0.6
django-extra-fields==3.0.2
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
pytz==2024.1
PyYAML==6.0.1
setuptools==70.1.1
sqlparse==0.5.0
uritemplate==4.1.1
